renamed cmake projects
