+++ lib/ziti_tunnel.c
@@ -27,11 +27,11 @@ limitations under the License.
 #include "lwip/raw.h"
 #include "lwip/timeouts.h"
 #include "netif_shim.h"
-#include "ziti/ziti_tunneler.h"
-#include "ziti_tunneler_priv.h"
+#include "ziti/ziti_tunnel.h"
+#include "ziti_tunnel_priv.h"
 #include "intercept.h"
-#include "tunneler_tcp.h"
-#include "tunneler_udp.h"
+#include "tunnel_tcp.h"
+#include "tunnel_udp.h"
 #include "uv.h"
 #include "ziti/ziti_log.h"
