+++ programs/ziti-edge-tunnel/CMakeLists.txt
@@ -1,5 +1,5 @@
 if(CMAKE_SYSTEM_NAME STREQUAL Darwin OR CMAKE_SYSTEM_NAME STREQUAL Linux)
-    project(ziti-tunnel-app)
+    project(ziti-edge-tunnel)
 
     if(CMAKE_SYSTEM_NAME STREQUAL Darwin)
         set(NETIF_DRIVER_SOURCE netif_driver/darwin/utun.c)
@@ -8,18 +8,18 @@ if(CMAKE_SYSTEM_NAME STREQUAL Darwin OR CMAKE_SYSTEM_NAME STREQUAL Linux)
         set(NETIF_DRIVER_SOURCE netif_driver/linux/tun.c)
     endif()
 
-    add_executable(ziti-tunnel-app ziti-tunnel-app.c ${NETIF_DRIVER_SOURCE} dnsmasq_manager.c)
-    set_property(TARGET ziti-tunnel-app PROPERTY C_STANDARD 11)
+    add_executable(ziti-edge-tunnel ziti-edge-tunnel.c ${NETIF_DRIVER_SOURCE} dnsmasq_manager.c)
+    set_property(TARGET ziti-edge-tunnel PROPERTY C_STANDARD 11)
 
 
-    target_compile_definitions(ziti-tunnel-app
+    target_compile_definitions(ziti-edge-tunnel
             PRIVATE GIT_VERSION=${GIT_VERSION}
             PRIVATE ZITI_LOG_PREFIX=${PROJECT_NAME}
             )
 
-    target_link_libraries(ziti-tunnel-app
-            PUBLIC ziti ziti_tunneler ziti_tunneler_cbs
-            PUBLIC lwipcore ${lwip_sys_lib} ziti_tunneler
+    target_link_libraries(ziti-edge-tunnel
+            PUBLIC ziti ziti-tunnel-sdk-c ziti-tunnel-cbs-c
+            PUBLIC lwipcore ${lwip_sys_lib}
             PUBLIC subcommand
             )
 endif()
